fix: PHPStan at level 5 in CI, and the bugs it found
PHPStan now runs in CI (pinned image, phpstan.neon at level 5, clean). What
it turned up:

- rowCount() after an UPDATE counted only rows whose values changed, so
  saving a habit (and most records) unchanged answered 404, and the
  workaround elsewhere, rowCount() >= 0, answered {"ok":true} for another
  account's exercise category or note tag. The connection now sets
  PDO::MYSQL_ATTR_FOUND_ROWS; those endpoints answer 404 for a row that is
  not the caller's, and renaming a tag tells "not yours" from "name taken".
- A project at 100% never got its "complete" insight: round() returns a
  float and the check was === 100.
- Response::json/redirect/abort always exit; they are declared never, which
  also removes dead "return;" lines after them.
- Exercise stats broke ties in no particular order, so equal counts swapped
  places between loads.
- Smaller: redundant checks, an always-true condition, a dead Response::view()
  that pointed at a missing template, an unused FinanceCategory::belongsTo(),
  GRANTABLE_ROLES moved to ProjectMember, bounds-checked DER parsing in
  WebPush, RateLimiter::hit() marked impure.

+++ models/Project.php
@@ -169,7 +169,8 @@ class Project
         $openCritical = (int)$stats['open_critical'];
         $overdue = (int)$stats['overdue'];
 
-        $percent = $total > 0 ? round(($done / $total) * 100) : 0;
+        // An int: compared with === 100 below, which a float from round() never equals.
+        $percent = $total > 0 ? (int)round(($done / $total) * 100) : 0;
         
         $status = 'healthy';
         $insight = '';
